Move java+py threading to its own file, start deephaven module later
